refactor : 메소드 추상화 일관성 유지하도록 setResponse로 로직 분리
